Test with larger matrix and process grid
Note: It is strange that when using Intel MPI and launching more than 16
processes on a single node, MPI_Win_flush_all() is blocked and the program
does not have any output. Launching 3*5 processes on a single node is OK,
launching 3*6 is not. Launching 1 process per node and using 16 nodes
works correctly.